requirements.txt
streamlit
sounddevice
numpy
webrtcvad
openai-whisper
torch
requests
